+++ test/e2e/ibc_transfer_e2e_test.go
@@ -1,12 +1,10 @@
 package e2e
 
 import (
-	"math"
 	"strings"
 	"time"
 
 	"github.com/babylonlabs-io/babylon/test/e2e/configurer"
-	"github.com/babylonlabs-io/babylon/test/e2e/initialization"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/suite"
 )
@@ -15,6 +13,8 @@ type IBCTransferTestSuite struct {
 	suite.Suite
 
 	configurer configurer.Configurer
+	addrA      string
+	addrB      string
 }
 
 func (s *IBCTransferTestSuite) SetupSuite() {
@@ -57,116 +57,165 @@ func getFirstIBCDenom(balance sdk.Coins) string {
 func (s *IBCTransferTestSuite) Test1IBCTransfer() {
 	denom := "ubbn"
 	amount := int64(1_000_000)
-	delta := float64(10000) // Tolerance to account for gas fees
 
 	transferCoin := sdk.NewInt64Coin(denom, amount)
 
 	bbnChainA := s.configurer.GetChainConfig(0)
 	bbnChainB := s.configurer.GetChainConfig(1)
 
-	babylonNodeA, err := bbnChainA.GetNodeAtIndex(2)
+	nA, err := bbnChainA.GetNodeAtIndex(2)
 	s.NoError(err)
-	babylonNodeB, err := bbnChainB.GetNodeAtIndex(2)
+	nB, err := bbnChainB.GetNodeAtIndex(2)
 	s.NoError(err)
-
-	val := initialization.ValidatorWalletName
 
 	// Check balance of val in chain-A (Node 3)
-	addrA := babylonNodeA.GetWallet(val)
-	balanceA, err := babylonNodeA.QueryBalances(addrA)
+	s.addrA = nA.KeysAdd("addr-A")
+	nA.BankSendFromNode(s.addrA, "10000000ubbn")
+
+	s.addrB = nB.KeysAdd("addr-B")
+	nB.BankSendFromNode(s.addrB, "10000000ubbn")
+
+	nB.WaitForNextBlock()
+	nA.WaitForNextBlock()
+
+	balanceBeforeSendAddrA, err := nA.QueryBalances(s.addrA)
 	s.Require().NoError(err)
 	// Confirm val on A has enough funds
-	s.Assert().GreaterOrEqual(balanceA.AmountOf(denom).Int64(), amount)
+	s.Assert().GreaterOrEqual(balanceBeforeSendAddrA.AmountOf(denom).Int64(), amount)
 
-	addrB := babylonNodeB.GetWallet(val)
-	balanceB, err := babylonNodeB.QueryBalances(addrB)
+	balanceBeforeSendAddrB, err := nB.QueryBalances(s.addrB)
 	s.Require().NoError(err)
 	// Only one denom in B
-	s.Require().Len(balanceB, 1)
+	s.Require().Len(balanceBeforeSendAddrB, 1)
 
 	// Send transfer from val in chain-A (Node 3) to val in chain-B (Node 3)
-	babylonNodeA.SendIBCTransfer(val, addrB, "transfer", transferCoin)
+	txHash := nA.SendIBCTransfer(s.addrA, s.addrB, "transfer", transferCoin)
+	nA.WaitForNextBlock()
 
+	_, txResp := nA.QueryTx(txHash)
+	txFeesPaid := txResp.AuthInfo.Fee.Amount
 	s.Require().Eventually(func() bool {
 		// Check that the transfer is successful.
 		// Amounts have been discounted from val in chain-A and added (as a wrapped denom) to val in chain-B
-		balanceA2, err := babylonNodeA.QueryBalances(addrA)
+		balanceAfterSendAddrA, err := nA.QueryBalances(s.addrA)
 		if err != nil {
+			s.T().Logf("failed to query balances: %s", err.Error())
 			return false
 		}
-		return math.Abs(float64(balanceA.Sub(transferCoin).AmountOf(denom).Int64()-
-			balanceA2.AmountOf(denom).Int64())) < delta
-	}, 10*time.Second, 1*time.Second, "Transfer was not successful")
+
+		expectedAmt := balanceBeforeSendAddrA.Sub(transferCoin).Sub(txFeesPaid...).String()
+		actualAmt := balanceAfterSendAddrA.String()
+
+		if !strings.EqualFold(expectedAmt, actualAmt) {
+			s.T().Logf(
+				"BalanceBeforeSendAddrA: %s; BalanceAfterSendAddrA: %s, txFees: %s, coinTransfer: %s",
+				balanceBeforeSendAddrA.String(), balanceAfterSendAddrA.String(), txFeesPaid.String(), transferCoin.String(),
+			)
+			return false
+		}
+
+		return true
+	}, 1*time.Minute, 1*time.Second, "Transfer was not successful")
 
 	s.Require().Eventually(func() bool {
-		balanceB2, err := babylonNodeB.QueryBalances(addrB)
+		balanceAfterSendAddrB, err := nB.QueryBalances(s.addrB)
 		if err != nil {
+			s.T().Logf("failed to query balances: %s", err.Error())
 			return false
 		}
 		// Check that there are now two denoms in B
-		if len(balanceB2) != 2 {
+		if len(balanceAfterSendAddrB) != 2 {
 			return false
 		}
-		denomB := getFirstIBCDenom(balanceB2)
-		// Check the balance of the IBC denom
-		return math.Abs(float64(balanceB2.AmountOf(denomB).Int64()-
-			transferCoin.Amount.Int64())) < delta
-	}, 10*time.Second, 1*time.Second, "Transfer was not successful")
+
+		denomB := getFirstIBCDenom(balanceAfterSendAddrB)
+		if !balanceAfterSendAddrB.AmountOf(denomB).Equal(transferCoin.Amount) {
+			s.T().Logf(
+				"BalanceBeforeSendAddrB: %s; BalanceBeforeSendAddrB: %s, txFees: %s, coinTransfer: %s",
+				balanceBeforeSendAddrB.String(), balanceAfterSendAddrB.String(), txFeesPaid.String(), transferCoin.String(),
+			)
+			return false
+		}
+
+		return true
+	}, 1*time.Minute, 1*time.Second, "Transfer was not successful")
 }
 
 func (s *IBCTransferTestSuite) Test2IBCTransferBack() {
 	nativeDenom := "ubbn"
-	delta := float64(10000) // Tolerance to account for gas fees
 
 	bbnChainA := s.configurer.GetChainConfig(0)
 	bbnChainB := s.configurer.GetChainConfig(1)
 
-	babylonNodeA, err := bbnChainA.GetNodeAtIndex(0)
+	nA, err := bbnChainA.GetNodeAtIndex(0)
 	s.NoError(err)
-	babylonNodeB, err := bbnChainB.GetNodeAtIndex(2)
+	nB, err := bbnChainB.GetNodeAtIndex(2)
 	s.NoError(err)
 
-	val := initialization.ValidatorWalletName
-
-	addrB := babylonNodeB.GetWallet(val)
-	balanceB, err := babylonNodeB.QueryBalances(addrB)
+	balanceBeforeSendBackB, err := nB.QueryBalances(s.addrB)
 	s.Require().NoError(err)
 	// Two denoms in B
-	s.Require().Len(balanceB, 2)
+	s.Require().Len(balanceBeforeSendBackB, 2)
 	// Look for the ugly IBC one
-	denom := getFirstIBCDenom(balanceB)
-	amount := balanceB.AmountOf(denom).Int64() - int64(delta) // have to pay gas fees
+	denom := getFirstIBCDenom(balanceBeforeSendBackB)
+	amount := balanceBeforeSendBackB.AmountOf(denom).Int64() // have to pay gas fees
 
 	transferCoin := sdk.NewInt64Coin(denom, amount)
 
 	// Send transfer from val in chain-B (Node 3) to val in chain-A (Node 1)
-	addrA := babylonNodeA.GetWallet(val)
-	balanceA, err := babylonNodeA.QueryBalances(addrA)
+	balanceBeforeReceivingSendBackA, err := nA.QueryBalances(s.addrA)
 	s.Require().NoError(err)
 
-	babylonNodeB.SendIBCTransfer(val, addrA, "transfer back", transferCoin)
+	txHash := nB.SendIBCTransfer(s.addrB, s.addrA, "transfer back", transferCoin)
+
+	nB.WaitForNextBlock()
+
+	_, txResp := nB.QueryTx(txHash)
+	txFeesPaid := txResp.AuthInfo.Fee.Amount
 
 	s.Require().Eventually(func() bool {
-		balanceB2, err := babylonNodeB.QueryBalances(addrB)
+		balanceAfterSendBackB, err := nB.QueryBalances(s.addrB)
 		if err != nil {
+			s.T().Logf("failed to query balances: %s", err.Error())
 			return false
 		}
-		return math.Abs(float64(balanceB.Sub(transferCoin).AmountOf(denom).Int64()-
-			balanceB2.AmountOf(denom).Int64())) < delta
-	}, 10*time.Second, 1*time.Second, "Transfer back was not successful")
+		expectedAmt := balanceBeforeSendBackB.Sub(transferCoin).Sub(txFeesPaid...).String()
+		actualAmt := balanceAfterSendBackB.String()
+
+		if !strings.EqualFold(expectedAmt, actualAmt) {
+			s.T().Logf(
+				"BalanceBeforeSendBackB: %s; BalanceAfterSendBackB: %s, txFees: %s, coinTransfer: %s",
+				balanceBeforeSendBackB.String(), balanceAfterSendBackB.String(), txFeesPaid.String(), transferCoin.String(),
+			)
+			return false
+		}
+
+		return true
+	}, 1*time.Minute, 1*time.Second, "Transfer back A was not successful")
 
 	nativeCoin := sdk.NewInt64Coin(nativeDenom, amount)
 	s.Require().Eventually(func() bool {
-		balanceA2, err := babylonNodeA.QueryBalances(addrA)
+		balanceAfterReceivingSendBackA, err := nA.QueryBalances(s.addrA)
 		if err != nil {
 			return false
 		}
 		// Check that there's still one denom in A
-		if len(balanceA2) != 1 {
+		if len(balanceAfterReceivingSendBackA) != 1 {
 			return false
 		}
+
+		expectedAmt := balanceBeforeReceivingSendBackA.Add(nativeCoin).String()
+		actualAmt := balanceAfterReceivingSendBackA.String()
+
 		// Check that the balance of the native denom has increased
-		return math.Abs(float64(balanceA.Add(nativeCoin).AmountOf(nativeDenom).Int64()-
-			balanceA2.AmountOf(nativeDenom).Int64())) < delta
-	}, 10*time.Second, 1*time.Second, "Transfer back was not successful")
+		if !strings.EqualFold(expectedAmt, actualAmt) {
+			s.T().Logf(
+				"BalanceBeforeReceivingSendBackA: %s; BalanceAfterReceivingSendBackA: %s, coinTransfer: %s",
+				balanceBeforeReceivingSendBackA.String(), balanceAfterReceivingSendBackA.String(), nativeCoin.String(),
+			)
+			return false
+		}
+
+		return true
+	}, 1*time.Minute, 1*time.Second, "Transfer back B was not successful")
 }